refactor: use map for TextSaver
