+++ app/src/espresso/java/xyz/lbres/trickcalculator/ui/main/testEqualsResult.kt
@@ -5,7 +5,7 @@ import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.espresso.matcher.ViewMatchers.withText
 import xyz.lbres.trickcalculator.R
-import xyz.lbres.trickcalculator.testutils.isEmptyString
+import xyz.lbres.trickcalculator.testutils.matchers.isEmptyString
 
 private val mainText = onView(withId(R.id.mainText))
